Fix days calculation on leave [half day] with [start,end,both]

+++ lib/features/Leave/data/models/leave_application_request_model.dart
@@ -28,7 +28,7 @@ class LeaveApplicationRequestModel extends LeaveApplicationRequestEntity {
       toDate: json['toDate'] as String?,
       toDateNp: json['toDateNp'] as String?,
       halfDayStatus: json['halfDayStatus'] as String?,
-      totalLeaveDays: json['totalLeaveDays'] as int?,
+      totalLeaveDays: json['totalLeaveDays'] as double?,
       extendedTotalLeaveDays: json['extendedTotalLeaveDays'] as int?,
       reason: json['reason'] as String?,
       extendedFromDate: json['extendedFromDate'] as String?,
